(chore) Restructure backend packages for better organization
