+++ backend/utils/settings.py
@@ -1,10 +1,6 @@
 from urllib.parse import urlparse
 
-from sqlalchemy import MetaData
-
-from utils.schema import get_table_properties, column_definition
 from utils.config import settings
-from utils.database import connect_database, get_unavailable_columns
 
 
 def get_all_sources():
@@ -37,34 +33,3 @@ def get_source_settings(source_label):
         return settings.DATABASES[requested_source[0]]
     elif requested_source[1] == "service":
         return getattr(settings, requested_source[2].upper())[requested_source[0]]
-
-
-async def get_source_database(source_settings, table_name=None, meta=None):
-    engine, conn = await connect_database(db_url=source_settings["db_url"])
-    if meta is None:
-        meta = MetaData(bind=engine)
-        meta.reflect()
-    unavailable_columns = get_unavailable_columns(source_settings=source_settings, meta=meta)
-    table_properties = get_table_properties(source_settings=source_settings, meta=meta)
-
-    if table_name and table_name in meta.tables.keys():
-        columns = [
-            column_definition(col, col_def) for col, col_def in meta.tables[table_name].columns.items()
-            if col not in unavailable_columns.get(table_name, [])
-        ]
-        conn.close()
-        return columns
-    else:
-        tables = sorted([
-            (
-                name, table_properties(name=name), [
-                    column_definition(col, col_def) for col, col_def in schema.columns.items()
-                    if col not in unavailable_columns.get(name, [])
-                ]
-            ) for name, schema in meta.tables.items()
-        ], key=lambda x: x[0])
-        conn.close()
-        return {
-            "columns": ["table_name", "properties", "columns"],
-            "rows": tables
-        }
